Add authorization hook #74
